(*) Fixed DSN handling in Migrate and Seed commands
