+++ tests/Phinx/Migration/Manager/EnvironmentTest.php
@@ -57,153 +57,6 @@ class EnvironmentTest extends TestCase
         $this->assertArrayHasKey('foo', $this->environment->getOptions());
     }
 
-    public function testConnectionOptionsCanBeSpecifiedWithDsn()
-    {
-        $dsn = 'pdomock://phinx:supersecret@my-database-host:1234/my_app_database';
-        $env = new Environment('testenv', ['dsn' => $dsn]);
-        $options = $env->getOptions();
-        $this->assertArrayHasKey('adapter', $options);
-        $this->assertSame('pdomock', $options['adapter']);
-        $this->assertArrayHasKey('user', $options);
-        $this->assertSame('phinx', $options['user']);
-        $this->assertArrayHasKey('pass', $options);
-        $this->assertSame('supersecret', $options['pass']);
-        $this->assertArrayHasKey('host', $options);
-        $this->assertSame('my-database-host', $options['host']);
-        $this->assertArrayHasKey('port', $options);
-        $this->assertEquals(1234, $options['port']);
-        $this->assertArrayHasKey('name', $options);
-        $this->assertSame('my_app_database', $options['name']);
-    }
-
-    public function testDsnOnlySetsSpecifiedOptions()
-    {
-        $dsn = 'pdomock://my-database-host/my_app_database';
-        $env = new Environment('testenv', ['dsn' => $dsn]);
-        $options = $env->getOptions();
-        $this->assertArrayNotHasKey('user', $options);
-        $this->assertArrayNotHasKey('pass', $options);
-        $this->assertArrayNotHasKey('port', $options);
-    }
-
-    public function testDsnGetsRemovedWhenAfterSuccessfulParsing()
-    {
-        $dsn = 'pdomock://phinx:supersecret@my-database-host:1234/my_app_database';
-        $env = new Environment('testenv', ['dsn' => $dsn]);
-        $this->assertArrayNotHasKey('dsn', $env->getOptions());
-    }
-
-    public function testOptionsAreLeftAsIsOnInvalidDsn()
-    {
-        $dsn = 'pdomock://phinx:supersecret@localhost:12badport34/db_name';
-        $env = new Environment('testenv', ['dsn' => $dsn]);
-        $options = $env->getOptions();
-        $this->assertArrayHasKey('dsn', $options);
-        $this->assertSame($dsn, $options['dsn']);
-    }
-
-    public function testDsnDoesNotOverrideSpecifiedOptions()
-    {
-        $dsn = 'pdomock://my-database-host:1234/my_web_database';
-        $env = new Environment('testenv', [
-            'user' => 'api_user',
-            'dsn' => $dsn,
-            'name' => 'my_api_database',
-        ]);
-        $options = $env->getOptions();
-        $this->assertArrayHasKey('user', $options);
-        $this->assertSame('api_user', $options['user']);
-        $this->assertArrayNotHasKey('pass', $options);
-        $this->assertArrayHasKey('name', $options);
-        $this->assertSame('my_api_database', $options['name']);
-    }
-
-    public function testNoModificationToOptionsOnInvalidDsn()
-    {
-        $dsn = 'pdomock://phinx:supersecret@localhost:12badport34/db_name';
-        $env = new Environment('testenv', [
-            'user' => 'api_user',
-            'dsn' => $dsn,
-            'name' => 'my_api_database',
-        ]);
-        $options = $env->getOptions();
-        $this->assertArrayHasKey('user', $options);
-        $this->assertSame('api_user', $options['user']);
-        $this->assertArrayHasKey('name', $options);
-        $this->assertSame('my_api_database', $options['name']);
-        $this->assertArrayHasKey('dsn', $options);
-    }
-
-    public function testDsnQueryProvidesAdditionalOptions()
-    {
-        $dsn = 'pdomock://phinx:supersecret@my-database-host:1234/my_app_database?charset=utf8&unrelated=thing&';
-        $env = new Environment('testenv', ['dsn' => $dsn]);
-        $options = $env->getOptions();
-        $this->assertArrayHasKey('charset', $options);
-        $this->assertSame('utf8', $options['charset']);
-        $this->assertArrayHasKey('unrelated', $options);
-        $this->assertSame('thing', $options['unrelated']);
-        $this->assertArrayNotHasKey('query', $options);
-    }
-
-    public function testDsnQueryDoesNotOverrideDsnParameters()
-    {
-        $dsn = 'pdomock://phinx:supersecret@my-database-host:1234/my_app_database?port=80&host=another-host';
-        $env = new Environment('testenv', ['dsn' => $dsn]);
-        $options = $env->getOptions();
-        $this->assertSame('my-database-host', $options['host']);
-        $this->assertEquals(1234, $options['port']);
-    }
-
-    public function dataProviderValidDsn()
-    {
-        return [
-            ['mysql://user:pass@host:1234/name?charset=utf8'],
-            ['postgres://user:pass@host/name?'],
-            ['mssql://user:@host:1234/name'],
-            ['sqlite3://user@host:1234/name'],
-            ['pdomock://host:1234/name'],
-            ['pdomock://user:pass@host/name'],
-            ['pdomock://host/name'],
-            // The RegEx that parses the DSN is purely to extract information
-            // from the string in the right order, it is not responsible for
-            // ensuring that the information is correct. Technically the
-            // following, whilst useless, is valid (validation will happen
-            // during connection):
-            ['£$%^&}{@>://$%*(*&^}{:?£}{@^>}"{$:1234/>@"}%{^>£}:/^'],
-            ['pdomock://user:pass@host/:1234/name'],
-            ['pdomock://user:pa:ss@host:1234/name'],
-        ];
-    }
-
-    public function dataProviderInvalidDsn()
-    {
-        return [
-            ['pdomock://user:pass@host:/name'],
-            ['pdomock://user:pass@host:1234/ '],
-            ['pdomock://user:pass@:1234/name'],
-            ['pdomock://:pass@host:1234/name'],
-            ['://user:pass@host:1234/name'],
-            ['pdomock:/user:p@ss@host:1234/name'],
-            ['pdomock://user:pass@host:/1234name'],
-            ['pdomock://user:pass@host:01234/name'],
-        ];
-    }
-
-    /** @dataProvider \Test\Phinx\Migration\Manager\EnvironmentTest::dataProviderValidDsn() */
-    public function testValidDsn($dsn)
-    {
-        $env = new Environment('testenv', ['dsn' => $dsn]);
-        $this->assertArrayNotHasKey('dsn', $env->getOptions());
-    }
-
-    /** @dataProvider \Test\Phinx\Migration\Manager\EnvironmentTest::dataProviderInvalidDsn() */
-    public function testInvalidDsn($dsn)
-    {
-        $env = new Environment('testenv', ['dsn' => $dsn]);
-        $this->assertArrayHasKey('dsn', $env->getOptions());
-    }
-
     /**
      * @expectedException \RuntimeException
      * @expectedExceptionMessage Adapter "fakeadapter" has not been registered
